fixed class Game and created classes Attack, GenerateLogs, Result
