+++ classes/classPlayer.js
